requirements.txt
streamlit
openai
Pillow
pytesseract
PyPDF2
pdf2image
tesseract-ocr
